requirements.txt
requests==2.32.3
beautifulsoup4==4.13.3
html5lib==1.1
pandas==2.2.3
uvicorn==0.34.0
fastapi==0.115.12
pytest==8.3.5
dotenv==0.9.9
PyJWT==2.10.1
python-multipart==0.0.20